refactor(web): extract all components in extra files, refactor imports
