--- apps/web/src/components/blocks/text.tsx
@@ -1,8 +1,0 @@
-/**
- * A simple text component.
- * @param {string} text
- * @returns {JSX.Element}
- */
-const Text = (text: string): JSX.Element => <p>{text}</p>;
-
-export default Text;
