--- apps/web/src/lib/helper.ts
@@ -1,3 +1,0 @@
-import * as R from 'ramda';
-
-export const pipeWhileNotNil = R.pipeWith((f, res) => (R.isNil(res) ? res : f(res)));
