+++ apps/web/src/pages/_app.tsx
@@ -1,4 +1,4 @@
-import '../styles/globals.css';
+import '../lib/styles/globals.css';
 import type {AppProps} from 'next/app';
 
 function MyApp({Component, pageProps}: AppProps) {
