+++ apps/web/src/pages/index.tsx
@@ -2,7 +2,7 @@ import type {NextPage} from 'next';
 import Head from 'next/head';
 import Image from 'next/image';
 import Link from 'next/link';
-import Layout from '../components/layout';
+import Layout from '../lib/components/Layout';
 import styles from '../styles/Home.module.css';
 
 const Home: NextPage = () => {
